apps/api/requirements.txt
flask
flask-sock
flask-cors
asyncssh
pyyaml
pytest